+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## x.x.x x-x-x
+### Payments
+* [Added] Support for CVC recollection PaymentSheet and PaymentSheet.FlowController (client-side confirmation)
+
 ## 23.21.2 2024-02-05
 ### Payments
 * [Changed] We now auto append `mandate_data` when using Klarna with a SetupIntent. If you are interested in using Klarna with SetupIntents you sign up for the beta [here](https://stripe.com/docs/payments/klarna/accept-a-payment).